rag-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
psycopg2-binary==2.9.10
pgvector==0.3.6
python-multipart==0.0.20
beautifulsoup4==4.12.3
pydantic-settings==2.7.1
sqlparse==0.5.3
redis==5.2.1
pytest==8.3.5
pytest-asyncio==0.25.3
fpdf2==2.8.2
